Reorganize parse tests
