+++ projects/adbquery/test/parser.test.js
@@ -1,13 +1,15 @@
 import * as parser from "../compiler/parser.js";
 
-test("invalid type", () => {
-    const data = {
-        Id: 10,
-    };
+describe("parse", () => {
+    test("invalid type", () => {
+        const data = {
+            Id: 10,
+        };
 
-    const parse = () => {
-        parser.parse(data);
-    };
+        const parse = () => {
+            parser.parse(data);
+        };
 
-    expect(parse).toThrow("Unknown token type 'number' of token 'Id'.");
+        expect(parse).toThrow("Unknown token type 'number' of token 'Id'.");
+    });
 });
